[Fix] Silencing one enum+float warning, introducing two new...
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@16437 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_gt2.cpp
@@ -442,7 +442,7 @@ static void TranslateGraoumfEffect(CSoundFile &sndFile, ModCommand &m, const uin
 				const double sampleDurationMiddleC = length / sampleRate;
 				const double wantedDuration = (param * currentSpeed * 2.5) / currentTempo.ToDouble();
 				const double ratio = sampleDurationMiddleC / wantedDuration;
-				const double note = Clamp(double(NOTE_MIDDLEC) + mpt::log2(ratio) * 12.0, NOTE_MIN, NOTE_MAX), frac = note - std::floor(note);
+				const double note = Clamp(double(NOTE_MIDDLEC) + mpt::log2(ratio) * 12.0, double(NOTE_MIN), double(NOTE_MAX)), frac = note - std::floor(note);
 
 				const bool hadNote = m.IsNote();
 				m.note = static_cast<ModCommand::NOTE>(note);
